requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
numpy==1.18.1
Pillow==7.0.0
plotly==4.2.0
requests==2.23.0
retrying==1.3.3
psutil==5.7.0
scipy==1.4.1
six==1.14.0
urllib3==1.25.8
wxPython==4.0.7.post2
# ======================== #
Django==3.0.6
djangorestframework==3.11.0